+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aave/core
 
+## 1.0.0-next.12
+
+### Minor Changes
+
+- 7ddec84: **feat:** implement `useSwapStatus` hook and makes `useUserSwaps` poll for status updates until terminal state.
+
+## 1.0.0-next.11
+
+### Patch Changes
+
+- 531b1f6: **feat:** dual-request strategy for `useTokenSwapQuote` with Fast and Accurate quotes
+
 ## 1.0.0-next.10
 
 ### Patch Changes